RavenDB-3784 - Refactoring RavenTraffic code adding fody usage to RavenTraffic and StorageExplorer modifying defaults.ps1 to build the RavenTraffic and StorageExplorer added RavenTraffic and StorageExplorer to zzz_RavenDB_Release.sln
