+++ Raven.Abstractions/Logging/LogHttpRequestStatsParams.cs
@@ -3,7 +3,7 @@ using System.Collections.Specialized;
 using System.Diagnostics;
 using System.Net.Http.Headers;
 
-namespace Raven.Database.Server
+namespace Raven.Abstractions.Logging
 {
 	public class LogHttpRequestStatsParams
 	{
