+++ Tools/Raven.Traffic/FodyWeavers.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Weavers>
+  <Costura IncludeDebugSymbols='false'>
+    <ExcludeAssemblies>
+      Raven.Abstractions
+      Raven.Database
+    </ExcludeAssemblies>
+  </Costura>
+</Weavers>
