+++ Tools/Raven.Traffic/TrafficRec.cs
@@ -13,8 +13,8 @@ using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
 using Raven.Abstractions.Data;
+using Raven.Abstractions.Logging;
 using Raven.Client;
-using Raven.Database.Server;
 using Raven.Imports.Newtonsoft.Json;
 using Raven.Json.Linq;
 using TrafficRecorder;
